+++ test/spy_pointer_test.cpp
@@ -23,26 +23,24 @@ using namespace mtl;
 struct SpiedObject
     : public enable_spying
 {
-    bool
-    spied_state_in_callback = false;
+    bool spied_state_in_callback = false;
     virtual void foo() {}
     virtual void on_spying_state_changed() { spied_state_in_callback = is_spied(); }
 };
 
-struct MockSpiedObject
+struct ChildObject
     : public SpiedObject
 {
     bool state = false;
-    MOCK_METHOD0(foo, void());
 };
 
 TEST(spying, can_dereference_spy)
 {
-    MockSpiedObject object;
-    spy_pointer<SpiedObject> spy(&object);
+    ChildObject obj;
+    spy_pointer<ChildObject> spy(&obj);
     ASSERT_TRUE(spy);
-    EXPECT_CALL(object, foo()).Times(1);
-    spy->foo();
+    spy->state = true;
+    EXPECT_TRUE(obj.state);
 }
 
 TEST(spying, can_detect_spied_object_destruction)
@@ -50,9 +48,12 @@ TEST(spying, can_detect_spied_object_destruction)
     spy_pointer<SpiedObject> spy;
     ASSERT_FALSE(spy);
     {
-        SpiedObject object;
-        spy.reset(&object);
+        SpiedObject obj;
+        EXPECT_FALSE(obj.is_spied());
+        
+        spy.reset(&obj);
         ASSERT_TRUE(spy);
+        EXPECT_TRUE(obj.is_spied());
     }
     ASSERT_FALSE(spy);
 }
@@ -61,10 +62,10 @@ TEST(spying, can_use_multiple_pointers)
 {
     spy_pointer<SpiedObject> spy1, spy2;
     {
-        SpiedObject object;
-        spy1.reset(&object);
+        SpiedObject obj;
+        spy1 = &obj;
         ASSERT_TRUE(spy1);
-        spy2.reset(&object);
+        spy2 = &obj;
         ASSERT_TRUE(spy2);
     }
     ASSERT_FALSE(spy1);
@@ -73,18 +74,19 @@ TEST(spying, can_use_multiple_pointers)
 
 TEST(spying, can_construct_a_copy)
 {
-    SpiedObject object;
-    spy_pointer<SpiedObject> spy1(&object);
+    SpiedObject obj;
+    spy_pointer<SpiedObject> spy1(&obj);
     spy_pointer<SpiedObject> spy2(spy1);
     EXPECT_TRUE(spy2);
     spy1.reset();
     EXPECT_TRUE(spy2);
+    EXPECT_TRUE(obj.is_spied());
 }
 
 TEST(spying, can_construct_by_move)
 {
-    SpiedObject object;
-    spy_pointer<SpiedObject> spy1(&object);
+    SpiedObject obj;
+    spy_pointer<SpiedObject> spy1(&obj);
     spy_pointer<SpiedObject> spy2(std::move(spy1));
     EXPECT_FALSE(spy1);
     EXPECT_TRUE(spy2);
@@ -142,27 +144,27 @@ TEST(spying, the_onli_first_and_last_spy_changes_spying_state)
 
 TEST(spying, can_construct_copy_of_casted_type)
 {
-    MockSpiedObject obj;
+    ChildObject obj;
     spy_pointer<SpiedObject> spy1(&obj);
-    spy_pointer<MockSpiedObject> spy2(spy1);
+    spy_pointer<ChildObject> spy2(spy1);
     spy2->state = true;
     EXPECT_TRUE(obj.state);
 }
 
 TEST(spying, can_construct_by_move_of_casted_type)
 {
-    MockSpiedObject obj;
+    ChildObject obj;
     spy_pointer<SpiedObject> spy1(&obj);
-    spy_pointer<MockSpiedObject> spy2(std::move(spy1));
+    spy_pointer<ChildObject> spy2(std::move(spy1));
     spy2->state = true;
     EXPECT_TRUE(obj.state);
 }
 
 TEST(spying, can_copy_of_casted_type)
 {
-    MockSpiedObject obj;
+    ChildObject obj;
     spy_pointer<SpiedObject> spy1(&obj);
-    spy_pointer<MockSpiedObject> spy2;
+    spy_pointer<ChildObject> spy2;
     spy2 = spy1;
     spy2->state = true;
     EXPECT_TRUE(obj.state);
@@ -170,11 +172,19 @@ TEST(spying, can_copy_of_casted_type)
 
 TEST(spying, can_move_casted_type)
 {
-    MockSpiedObject obj;
+    ChildObject obj;
     spy_pointer<SpiedObject> spy1(&obj);
-    spy_pointer<MockSpiedObject> spy2;
+    spy_pointer<ChildObject> spy2;
     spy2 = std::move(spy1);
     spy2->state = true;
+    EXPECT_TRUE(obj.state);
+}
+
+TEST(spying, can_cast_type)
+{
+    ChildObject obj;
+    spy_pointer<SpiedObject> spy(&obj);
+    static_cast<spy_pointer<ChildObject>>(spy)->state = true;
     EXPECT_TRUE(obj.state);
 }